requirements.txt
adal==1.2.3
azure-datalake-store==0.0.48
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
idna==2.9
numpy==1.18.5
pandas==0.25.3
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.15.0
SoundFile==0.10.3.post1
urllib3==1.25.9
